string_is_email and email_like validators splitted
