+++ validators/string_is_email_like_test.go
@@ -4,11 +4,10 @@ import (
 	"testing"
 
 	"github.com/s3rj1k/validator"
-
 	"github.com/stretchr/testify/require"
 )
 
-func Test_StringIsEmail(t *testing.T) {
+func Test_StringIsEmailLike(t *testing.T) {
 
 	r := require.New(t)
 
@@ -25,6 +24,7 @@ func Test_StringIsEmail(t *testing.T) {
 		{"foo@bar.中文网", true},
 		{"invalidemail@", false},
 		{"invalid.com", false},
+		{"@", false},
 		{"@invalid.com", false},
 		{"test|123@m端ller.com", true},
 		{"hans@m端ller.com", true},
@@ -34,18 +34,24 @@ func Test_StringIsEmail(t *testing.T) {
 	}
 
 	for _, test := range tests {
-		v := &StringIsEmail{Name: "email", Field: test.email}
+		v := &StringIsEmailLike{Name: "email", Field: test.email}
 		e := validator.NewErrors()
 		v.Validate(e)
-		r.Equal(test.valid, !e.HasAny())
+		r.Equal(test.valid, !e.HasAny(), test.email)
 		if !test.valid {
-			r.Equal([]string{StringIsEmailError(v)}, e.Get("email"))
+			r.Equal(e.Get("email"), []string{StringIsEmailLikeError(v)})
 		}
 	}
 
-	v := &StringIsEmail{Name: "email", Field: ""}
+	v := &StringIsEmailLike{Name: "email", Field: "foo@bar"}
 	e := validator.NewErrors()
 	v.Validate(e)
 	r.Equal(e.Count(), 1)
-	r.Equal([]string{StringIsEmailError(v)}, e.Get("email"))
+	r.Equal(e.Get("email"), []string{StringIsEmailLikeError(v)})
+
+	v = &StringIsEmailLike{Name: "email", Field: ""}
+	e = validator.NewErrors()
+	v.Validate(e)
+	r.Equal(e.Count(), 1)
+	r.Equal(e.Get("email"), []string{StringIsEmailLikeError(v)})
 }
